Name fixed
I'm bad at names

errors still remain???
