+++ src/main/java/competition/subsystems/humanLoadRamp/commands/HumanLoadRampExtendCommand.java
@@ -1,20 +1,19 @@
-package competition.subsystems.humanLoadRamp.Commands;
+package competition.subsystems.humanLoadRamp.commands;
 
 import competition.subsystems.humanLoadRamp.HumanLoadRampSubsystem;
 import xbot.common.command.BaseCommand;
 
 import javax.inject.Inject;
 
-public class HumanLoadRampRetractCommand extends BaseCommand {
+public class HumanLoadRampExtendCommand extends BaseCommand {
     HumanLoadRampSubsystem ramp;
 
     @Inject
-    public HumanLoadRampRetractCommand (HumanLoadRampSubsystem humanLoadRampSubsystem){
+    public HumanLoadRampExtendCommand (HumanLoadRampSubsystem humanLoadRampSubsystem){
         ramp = humanLoadRampSubsystem;
         this.addRequirements(ramp);
-
     }
     public void initialize(){
-        ramp.retract();
+        ramp.extend();
     }
 }
